@@ -10,20 +10,20 @@
   "author": "",
   "license": "MIT",
   "devDependencies": {
-    "@angular/animations": "^4.4.6",
-    "@angular/platform-browser": "^4.4.6",
-    "angular2-notifications": "^0.7.8",
+    "@angular/animations": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "angular2-notifications": "^0.9.7",
     "prismjs": "^1.13.0",
-    "@angular/common": "^4.4.4",
-    "@angular/compiler": "^4.4.4",
-    "@angular/compiler-cli": "^4.4.4",
-    "@angular/core": "^4.4.4",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/compiler-cli": "^5.0.0",
+    "@angular/core": "^5.0.0",
     "@types/node": "^8.0.54",
     "gulp": "^3.9.1",
     "gulp-inline-ng2-template": "^4.0.0",
-    "rxjs": "^5.0.1",
-    "typescript": "^2.4.2",
+    "rxjs": "^5.5.2",
+    "typescript": "~2.4.2",
     "node-sass": "^4.7.2",
-    "zone.js": "^0.8.4"
+    "zone.js": "^0.8.14"
   }
 }
