@@ -3,7 +3,7 @@ import { ActionAlertComponent } from './action-alert/action-alert.component';
 import { NotificationItemComponent } from './notifications/notification-item/notification-item.component';
 import { NotificationQueueComponent } from './notifications/notification-queue/notification-queue.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { NotificationsService } from 'angular2-notifications/dist';
+import { NotificationsService } from 'angular2-notifications';
 import { IconComponent } from './icon/icon.component';
 import { CommonModule } from '@angular/common';
 import { CodeHighlightDirective } from './code-highlight/code-highlight.directive';
