@@ -1,6 +1,7 @@
-import {Component, ViewEncapsulation, OnInit, Input} from '@angular/core';
+import { Component, ViewEncapsulation, OnInit, Input, ChangeDetectorRef, OnDestroy } from '@angular/core';
 
 import {SimpleNotificationsComponent, NotificationsService, Notification} from 'angular2-notifications';
+import { Subscription } from 'rxjs/Subscription';
 
 const SELECTOR: string = 'notification-queue';
 
@@ -11,15 +12,20 @@ const SELECTOR: string = 'notification-queue';
   encapsulation: ViewEncapsulation.None
 })
 
-export class NotificationQueueComponent extends SimpleNotificationsComponent implements OnInit {
+export class NotificationQueueComponent extends SimpleNotificationsComponent implements OnInit, OnDestroy {
 
-  @Input() service: NotificationsService;
+  @Input('service') localService: NotificationsService;
 
-  constructor(private _notificationsService: NotificationsService) {
-    super(_notificationsService);
+  private localListener: Subscription;
+
+  constructor(
+    private _notificationsService: NotificationsService,
+    private _cdr: ChangeDetectorRef
+  ) {
+    super(_notificationsService, _cdr);
 
     super.attachChanges({
-      position: ["top", "left"],
+      position: ['top', 'left'],
       animate: 'fromTop',
       timeOut: 10000,
       showProgressBar: true,
@@ -30,9 +36,9 @@ export class NotificationQueueComponent extends SimpleNotificationsComponent imp
   }
 
   ngOnInit() {
-    if (this.service) {
-      // This is indentical to super.ngOnInit() but with non-global service
-      this.service.getChangeEmitter()
+    if (this.localService) {
+      // This is identical to super.ngOnInit() but with non-global service
+      this.localListener = this.localService.emitter
         .subscribe(item => {
           switch (item.command) {
             case 'cleanAll':
@@ -47,7 +53,7 @@ export class NotificationQueueComponent extends SimpleNotificationsComponent imp
                 this.add(item.notification!)
               } else {
                 this.defaultBehavior(item)
-              };
+              }
               break;
 
             default:
@@ -62,7 +68,13 @@ export class NotificationQueueComponent extends SimpleNotificationsComponent imp
   }
 
   onItemClose(item: Notification) {
-    const service = this.service || this._notificationsService;
+    const service = this.localService || this._notificationsService;
     service.set(item, false);
+  }
+
+  ngOnDestroy(): void {
+    if (this.localListener) {
+      this.localListener.unsubscribe();
+    }
   }
 }
