@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import {NotificationsService} from 'angular2-notifications/dist';
+import { NotificationsService } from 'angular2-notifications';
 
 @Component({
   selector: 'app-root',
