@@ -3,6 +3,7 @@ import { NgModule } from '@angular/core';
 import { GeneralModule} from 'xynga-general';
 
 import { AppComponent } from './app.component';
+import { NotificationsService } from 'angular2-notifications';
 
 
 @NgModule({
@@ -13,7 +14,7 @@ import { AppComponent } from './app.component';
     BrowserModule,
     GeneralModule
   ],
-  providers: [],
+  providers: [NotificationsService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
